Add Automated Labelling + New Label for PRs (#17)
Keyword `Type:` from PR Body can be used to set an aspect label to
the PR which will then be used for Release Change-logging. Otherwise
by setting the `Fixes #{Number}` keyword on the first line of the PR
Body, the bot will assign an aspect label that is equivalent
to that placed on the issue.

Also fix implemented for PR Labelling bug. If PR is Opened as a Draft
the Ongoing Label will be appended instead of ToReview.